do not bundle a library whose classes are already provideded by core
the activation framework classes are already provided by core.

Prior to jenkins-2.235 they where patched 1.1.1 versions but since
https://github.com/jenkinsci/jenkins/commit/abb131073ac06b2129cbda860c53d4720402a47f
it is the 1.2.1 version transitively from jakarta.mail

As the plugin is neither using plugin first classloader nor the
mask-classes approach this is just bloat - so remote it.

bumps the parent version and the minimum jenkins baseline to be > 2.235
to ensure we do get at least a 1.2.x version